+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 **Hearthstone**:
 - Fixed the secret tracker showing the wrong secrets when The Origin Stone cast a secret from a Discover.
 
+**Plugins**:
+- Fixed a plugin compatibility issue.
+
 ## **Release v1.55.10 - 2026-08-14**
 **Battlegrounds**:
 - Fixed an issue with the instrumentation we use to fix incorrect combat odds.