+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## **Unreleased**
+## **Release v1.55.11 - 2026-08-15**
 **Hearthstone**:
 - Fixed the secret tracker showing the wrong secrets when The Origin Stone cast a secret from a Discover.